Release v0.9.1

+++ tests/integration/conftest.py
@@ -5,6 +5,7 @@ import boto3
 import uuid
 import json
 from botocore.exceptions import ClientError
+from botocore.config import Config
 from requests_aws4auth import AWS4Auth
 
 from phebee.utils.aws import get_client
@@ -14,6 +15,15 @@ from step_function_utils import start_step_function, wait_for_step_function_comp
 from s3_utils import delete_s3_prefix
 from update_source_utils import update_source
 from constants import PROJECT_CONFIGS
+
+
+def pytest_addoption(parser):
+    parser.addoption(
+        "--skip-ontology-updates",
+        action="store_true",
+        default=False,
+        help="Skip slow HPO and MONDO graph updates during tests.",
+    )
 
 
 @pytest.fixture(scope="session")
@@ -28,11 +38,13 @@ def aws_session(request, profile_name):
     # Create a client for each service, so that they are ready to use.
     # If any other clients are needed, add them here so they are instantiated as well.
     get_client("s3", profile_name)
-    get_client("lambda", profile_name)
     get_client("cloudformation", profile_name)
     get_client("stepfunctions", profile_name)
     get_client("dynamodb", profile_name)
 
+    lambda_config = Config(read_timeout=900, connect_timeout=900, retries={"max_attempts": 0})
+    get_client("lambda", profile_name, lambda_config)
+    
     session = boto3.Session(profile_name=profile_name)
 
     return session
@@ -83,11 +95,7 @@ def cloudformation_stack(request, aws_session, profile_name):
     existing_stack = request.config.getoption("--existing-stack")
     config_env = request.config.getoption("--config-env")
     stack_uuid = str(uuid.uuid4())[-3:]
-    stack_name = (
-        f"phebee-integration-test-{stack_uuid}"
-        if not existing_stack
-        else existing_stack
-    )
+    stack_name = f"phebee-it-{stack_uuid}" if not existing_stack else existing_stack
 
     template_path = "template.yaml"
 
@@ -157,8 +165,18 @@ def cloudformation_stack(request, aws_session, profile_name):
     if not existing_stack:
         # Tear down the stack after tests
         try:
+            teardown_args = [
+                "sam",
+                "delete",
+                "--stack-name",
+                stack_name,
+                "--no-prompts",
+            ]
+            if profile_name:
+                teardown_args.append("--profile")
+                teardown_args.append(profile_name)
             subprocess.run(
-                ["sam", "delete", "--stack-name", stack_name, "--no-prompts"],
+                teardown_args,
                 check=True,
             )
         except subprocess.CalledProcessError as e:
@@ -239,19 +257,43 @@ def physical_resources(aws_session, cloudformation_stack):
 
 
 @pytest.fixture(scope="session")
-def update_hpo(cloudformation_stack, physical_resources):
-    test_start_time = update_source(cloudformation_stack, "hpo", "UpdateHPOSFNArn")
-    yield test_start_time
+def update_hpo(request, cloudformation_stack, physical_resources):
+    if request.config.getoption("--skip-ontology-updates"):
+        print("Skipping HPO update as requested via --skip-ontology-updates parameter")
+        yield
+    else:
+        test_start_time = update_source(cloudformation_stack, "hpo", "UpdateHPOSFNArn")
+        yield test_start_time
 
-    delete_s3_prefix(physical_resources["PheBeeBucket"], "sources/hpo")
+        delete_s3_prefix(physical_resources["PheBeeBucket"], "sources/hpo")
 
 
 @pytest.fixture(scope="session")
-def update_mondo(cloudformation_stack, physical_resources):
-    test_start_time = update_source(cloudformation_stack, "mondo", "UpdateMONDOSFNArn")
-    yield test_start_time
+def update_mondo(request, cloudformation_stack, physical_resources):
+    if request.config.getoption("--skip-ontology-updates"):
+        print(
+            "Skipping Mondo update as requested via --skip-ontology-updates parameter"
+        )
+        yield
+    else:
+        test_start_time = update_source(
+            cloudformation_stack, "mondo", "UpdateMondoSFNArn"
+        )
+        yield test_start_time
 
-    delete_s3_prefix(physical_resources["PheBeeBucket"], "sources/mondo")
+        delete_s3_prefix(physical_resources["PheBeeBucket"], "sources/mondo")
+
+
+@pytest.fixture(scope="session")
+def update_eco(request, cloudformation_stack, physical_resources):
+    if request.config.getoption("--skip-ontology-updates"):
+        print("Skipping ECO update as requested via --skip-ontology-updates parameter")
+        yield
+    else:
+        test_start_time = update_source(cloudformation_stack, "eco", "UpdateECOSFNArn")
+        yield test_start_time
+
+        delete_s3_prefix(physical_resources["PheBeeBucket"], "sources/eco")
 
 
 @pytest.fixture
@@ -267,11 +309,13 @@ def test_project_id(cloudformation_stack):
         Payload=json.dumps({"body": json.dumps(payload)}),
     )
 
-    body = json.loads(json.loads(create_response["Payload"].read().decode("utf-8"))["body"])
+    body = json.loads(
+        json.loads(create_response["Payload"].read().decode("utf-8"))["body"]
+    )
     assert create_response["StatusCode"] == 200, f"Invoke failed: {body}"
     assert body.get("project_created") is True
 
-    yield project_id
+    yield body.get("project_id")
 
     # Teardown
     delete_response = lambda_client.invoke(
@@ -285,25 +329,116 @@ def test_project_id(cloudformation_stack):
         print(f"WARNING: Teardown failed: {result}")
 
 
-# Fixture to upload phenopacket test data to S3
-@pytest.fixture(scope="session")
-def upload_phenopacket_s3(request, cloudformation_stack, physical_resources):
-    config = PROJECT_CONFIGS[request.param]
-    s3_client = get_client("s3")
+@pytest.fixture
+def create_test_subject(physical_resources):
+    lambda_client = get_client("lambda")
+    created_subjects = []
+    project_id = f"test-proj-{uuid.uuid4().hex[:6]}"
 
-    try:
-        zip_key = f"{request.param}/{os.path.basename(config['ZIP_PATH'])}"
-        with open(config["ZIP_PATH"], "rb") as data:
-            s3_client.put_object(
-                Bucket=physical_resources["PheBeeBucket"], Key=zip_key, Body=data
-            )
-        yield zip_key
-    except ClientError as e:
-        pytest.fail(f"Failed to upload ZIP to S3: {e}")
+    # Create project
+    project_response = lambda_client.invoke(
+        FunctionName=physical_resources["CreateProjectFunction"],
+        Payload=json.dumps(
+            {
+                "body": json.dumps(
+                    {"project_id": project_id, "project_label": project_id}
+                )
+            }
+        ).encode("utf-8"),
+        InvocationType="RequestResponse",
+    )
+    project_body = json.loads(
+        json.loads(project_response["Payload"].read().decode("utf-8"))["body"]
+    )
+    print(project_body)
 
-    # Cleanup after tests
-    finally:
-        s3_client.delete_object(Bucket=physical_resources["PheBeeBucket"], Key=zip_key)
+    def _make_subject():
+        project_subject_id = f"test-subj-{uuid.uuid4().hex[:6]}"
+        payload = {"project_id": project_id, "project_subject_id": project_subject_id}
+
+        response = lambda_client.invoke(
+            FunctionName=physical_resources["CreateSubjectFunction"],
+            Payload=json.dumps({"body": json.dumps(payload)}).encode("utf-8"),
+            InvocationType="RequestResponse",
+        )
+
+        body = json.loads(response["Payload"].read())
+
+        print(body)
+
+        subject_data = json.loads(body["body"])["subject"]
+
+        created_subjects.append(
+            {"project_id": project_id, "project_subject_id": project_subject_id}
+        )
+
+        return subject_data
+
+    yield _make_subject
+
+    # Cleanup
+    for subj in created_subjects:
+        lambda_client.invoke(
+            FunctionName=physical_resources["RemoveSubjectFunction"],
+            Payload=json.dumps(
+                {
+                    "body": json.dumps(
+                        {
+                            "project_id": subj["project_id"],
+                            "project_subject_id": subj["project_subject_id"],
+                        }
+                    )
+                }
+            ).encode("utf-8"),
+            InvocationType="RequestResponse",
+        )
+
+    lambda_client.invoke(
+        FunctionName=physical_resources["RemoveProjectFunction"],
+        Payload=json.dumps({"body": json.dumps({"project_id": project_id})}).encode(
+            "utf-8"
+        ),
+        InvocationType="RequestResponse",
+    )
+
+
+@pytest.fixture
+def create_test_encounter_iri(create_test_subject, physical_resources):
+    lambda_client = get_client("lambda")
+    created_encounters = []
+
+    def _make_encounter():
+        subject_iri = create_test_subject()["iri"]
+        encounter_id = f"test-enc-{uuid.uuid4().hex[:6]}"
+        payload = {"subject_iri": subject_iri, "encounter_id": encounter_id}
+
+        print(f"payload: {payload}")
+
+        response = lambda_client.invoke(
+            FunctionName=physical_resources["CreateEncounterFunction"],
+            Payload=json.dumps({"body": json.dumps(payload)}).encode("utf-8"),
+            InvocationType="RequestResponse",
+        )
+
+        body = json.loads(json.loads(response["Payload"].read())["body"])
+
+        print(f"body: {body}")
+
+        encounter_iri = body["encounter_iri"]
+
+        created_encounters.append((subject_iri, encounter_id))
+        return encounter_iri
+
+    yield _make_encounter
+
+    # Cleanup
+    for subject_iri, encounter_id in created_encounters:
+        payload = {"subject_iri": subject_iri, "encounter_id": encounter_id}
+        lambda_client.invoke(
+            FunctionName=physical_resources["RemoveEncounterFunction"],
+            Payload=json.dumps({"body": json.dumps(payload)}).encode("utf-8"),
+            InvocationType="RequestResponse",
+        )
 
 
 # Fixture to create the test project
@@ -345,47 +480,6 @@ def create_test_project(request, cloudformation_stack, physical_resources):
         assert result["statusCode"] == 200, (
             f"Failed to remove the {config['PROJECT_ID']} project"
         )
-
-
-# Fixture to import phenopacket data
-@pytest.fixture(scope="session")
-def import_phenopacket(
-    request,
-    upload_phenopacket_s3,
-    create_test_project,
-    physical_resources,
-    update_hpo,
-    update_mondo,
-):
-    config = PROJECT_CONFIGS[request.param]
-    s3_key = upload_phenopacket_s3
-
-    # Start the step function for import
-    bucket_arn = physical_resources["PheBeeBucket"]
-    s3_path = f"s3://{bucket_arn}/{s3_key}"
-    output_s3_path = f"s3://{bucket_arn}/test_import_packets.json"
-    step_function_input = json.dumps(
-        {
-            "project_id": config["PROJECT_ID"],
-            "s3_path": s3_path,
-            "output_s3_path": output_s3_path,
-        }
-    )
-
-    try:
-        # Start the step function and wait for completion
-        execution_arn = start_step_function(
-            physical_resources["ImportPhenopacketsSFN"],
-            step_function_input,
-        )
-        result = wait_for_step_function_completion(execution_arn, return_response=True)
-        yield result
-    except ClientError as e:
-        pytest.fail(f"Step function failed: {e}")
-
-    finally:
-        s3_client = get_client("s3")
-        s3_client.delete_object(Bucket=bucket_arn, Key="test_import_packets.json")
 
 
 def pytest_collection_modifyitems(session, config, items):
